[#52] refactor: getUser, isIdDuplicated 수정

+++ src/main/java/com/festa/dto/MemberDTO.java
@@ -19,7 +19,7 @@ import javax.validation.constraints.Pattern;
 @Builder
 public class MemberDTO {
 
-    int userNo;
+    long userNo;
 
     @NotBlank(message = "아이디를 입력해주세요")
     long userId;
